Pulled the fields from summary and contact information into Resume object
