--- src/main/java/com/garbuz/resume/entity/ContactInformation.java
@@ -1,149 +1,0 @@
-package com.garbuz.resume.entity;
-
-import java.util.Objects;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "ContactInformation")
-@JsonIgnoreProperties({"resume"})
-public class ContactInformation {
-	@Id
-	@Column(name = "id")
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
-	@Column(name = "email")
-	private String email;
-	@Column(name = "phone")
-	private String phone;
-	@Column(name = "address")
-	private String address;
-	@Column(name = "city")
-	private String city;
-	@Column(name = "state", length = 2)
-	private String state;
-	@Column(name = "zip", length = 5)
-	private String zip;
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "resume_id")
-	private Resume resume;
-
-	public ContactInformation() {
-	}
-
-
-	public ContactInformation(Resume resume, String email, String phone, String address, String city, String state, String zip) {
-		this.email = email;
-		this.phone = phone;
-		this.address = address;
-		this.city = city;
-		this.state = state;
-		this.zip = zip;
-		this.resume = resume;
-	}
-
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getPhone() {
-		return phone;
-	}
-
-	public void setPhone(String phone) {
-		this.phone = phone;
-	}
-
-	public String getAddress() {
-		return address;
-	}
-
-	public void setAddress(String address) {
-		this.address = address;
-	}
-
-	public String getCity() {
-		return city;
-	}
-
-	public void setCity(String city) {
-		this.city = city;
-	}
-
-	public String getState() {
-		return state;
-	}
-
-	public void setState(String state) {
-		this.state = state;
-	}
-
-	public String getZip() {
-		return zip;
-	}
-
-	public void setZip(String zip) {
-		this.zip = zip;
-	}
-
-	public Resume getResume() {
-		return resume;
-	}
-
-	public void setResume(Resume resume) {
-		this.resume = resume;
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(address, city, email, id, phone, state, zip);
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		ContactInformation other = (ContactInformation) obj;
-		return Objects.equals(address, other.address) && Objects.equals(city, other.city)
-				&& Objects.equals(email, other.email) 
-				&& Objects.equals(id, other.id) 
-				&& Objects.equals(phone, other.phone) && Objects.equals(state, other.state)
-				&& Objects.equals(zip, other.zip);
-	}
-
-	@Override
-	public String toString() {
-		return "ContactInformation [id=" + id + ", email="
-				+ email + ", phone=" + phone + ", address=" + address + ", city=" + city + ", state=" + state + ", zip="
-				+ zip + "]";
-	}
-
-
-}
